♻️ 라우터 분리 후 lasy, suspense 추가
- 초기 렌더링을 빠르게 불러오기 위해 각 페이지 별로 lasy로딩
- suspense는 기존에 존재하던 로딩컴포넌트 사용
